Moved static state manager to plugin top level
